implemented a solution that build from changed files only
It operates by reading the filepaths that have been changed
during this commit. Then it takes the blobs associated with
those commits to search for issue data or changes to issue
data. For the issues in files that have not changed, the
issues are brought forward to this current commit.